Re-factored health module
