+++ adapters/driven/module.go
@@ -3,6 +3,7 @@ package driven
 import (
 	"github.com/NicklasWallgren/go-template/adapters/driven/env"
 	"github.com/NicklasWallgren/go-template/adapters/driven/health"
+	"github.com/NicklasWallgren/go-template/adapters/driven/health/checkers"
 	"github.com/NicklasWallgren/go-template/adapters/driven/logger"
 	"github.com/NicklasWallgren/go-template/adapters/driven/persistence"
 	"github.com/NicklasWallgren/go-template/adapters/driven/persistence/migration"
@@ -31,8 +32,8 @@ var consumerManager = fx.Provide(
 var publishers = fx.Provide(rabbitmq.NewRabbitMQPublisher)
 
 var healthCheckers = fx.Provide(
-	fx.Annotate(health.NewDBHealthChecker, fx.ResultTags(`group:"checkers"`)),
-	fx.Annotate(health.NewRabbitMQHealthChecker, fx.ResultTags(`group:"checkers"`)),
+	fx.Annotate(checkers.NewDBHealthChecker, fx.ResultTags(`group:"checkers"`)),
+	fx.Annotate(checkers.NewRabbitMQHealthChecker, fx.ResultTags(`group:"checkers"`)),
 )
 
 var healthCheckerManager = fx.Provide(
